// types/three/three-core.d.ts
export class Camera extends Object3D {
    /**
     * This constructor sets following properties to the correct type: matrixWorldInverse, projectionMatrix and projectionMatrixInverse.
     */
    constructor();

    /**
     * This is the inverse of matrixWorld. MatrixWorld contains the Matrix which has the world transform of the Camera.
     */
    matrixWorldInverse: Matrix4;

    /**
     * This is the matrix which contains the projection.
     */
    projectionMatrix: Matrix4;

    getWorldDirection(optionalTarget?: Vector3): Vector3;

    /**
     * This make the camera look at the vector position in local space.
     * @param vector point to look at
     */
    lookAt(vector: Vector3): void;
}
// ...
export class BufferAttribute {
    constructor(array: ArrayLike<number>, itemSize: number, normalized?: boolean); // array parameter should be TypedArray.

    uuid: string;
    array: ArrayLike<number>;
    itemSize: number;
    dynamic: boolean;
    updateRange: {offset: number, count: number};
    version: number;
    normalized: boolean;
    needsUpdate: boolean;
    count: number;

    setDynamic(dynamic: boolean): BufferAttribute;
    clone(): this;
    copy(source: this): this;
    copyAt(index1: number, attribute: BufferAttribute, index2: number): BufferAttribute;
    copyArray(array: ArrayLike<number>): BufferAttribute;
    copyColorsArray(colors: {r: number, g: number, b: number}[]): BufferAttribute;
    copyIndicesArray(indices: {a: number, b: number, c: number}[]): BufferAttribute;
    copyVector2sArray(vectors: {x: number, y: number}[]): BufferAttribute;
    copyVector3sArray(vectors: {x: number, y: number, z: number}[]): BufferAttribute;
    copyVector4sArray(vectors: {x: number, y: number, z: number, w: number}[]): BufferAttribute;
    set(value: ArrayLike<number>, offset?: number): BufferAttribute;
    getX(index: number): number;
    setX(index: number, x: number): BufferAttribute;
    getY(index: number): number;
    setY(index: number, y: number): BufferAttribute;
    getZ(index: number): number;
    setZ(index: number, z: number): BufferAttribute;
    getW(index: number): number;
    setW(index: number, z: number): BufferAttribute;
    setXY(index: number, x: number, y: number): BufferAttribute;
    setXYZ(index: number, x: number, y: number, z: number): BufferAttribute;
    setXYZW(index: number, x: number, y: number, z: number, w: number): BufferAttribute;
    /**
     * @deprecated Use count instead.
     */
    length: number;
}
// ...
export class Font {
    constructor(jsondata: string);

    data: string;

    generateShapes(text: string, size: number, divisions: number): any[];
}
// ...
export class Euler {
    constructor(x?: number, y?: number, z?: number, order?: string);

    x: number;
    y: number;
    z: number;
    order: string;
    onChangeCallback: Function;

    set(x: number, y: number, z: number, order?: string): Euler;
    clone(): this;
    copy(euler: this): this;
    setFromRotationMatrix(m: Matrix4, order?: string, update?: boolean): Euler;
    setFromQuaternion(q: Quaternion, order?: string, update?: boolean): Euler;
    setFromVector3( v: Vector3, order?: string ): Euler;
    reorder(newOrder: string): Euler;
    equals(euler: Euler): boolean;
    fromArray(xyzo: any[]): Euler;
    toArray(array?: number[], offset?: number): number[];
    toVector3(optionalResult?: Vector3): Vector3;
    onChange(callback: Function): void;

    static RotationOrders: string[];
    static DefaultOrder: string;
}
// ...
export class Matrix3 implements Matrix {
    /**
     * Creates an identity matrix.
     */
    constructor();

    /**
     * Float32Array with matrix values.
     */
    elements: Float32Array;

    set(n11: number, n12: number, n13: number, n21: number, n22: number, n23: number, n31: number, n32: number, n33: number): Matrix3;
    identity(): Matrix3;
    clone(): this;
    copy(m: this): this;
    setFromMatrix4(m: Matrix4): Matrix3;
    applyToBuffer(buffer: BufferAttribute, offset?: number, length?: number): BufferAttribute;
    multiplyScalar(s: number): Matrix3;
    determinant(): number;
    getInverse(matrix: Matrix3, throwOnDegenerate?: boolean): Matrix3;

    /**
     * Transposes this matrix in place.
     */
    transpose(): Matrix3;
    getNormalMatrix(matrix4: Matrix4): Matrix3;

    /**
     * Transposes this matrix into the supplied array r, and returns itself.
     */
    transposeIntoArray(r: number[]): number[];
    fromArray(array: number[], offset?: number): Matrix3;
    toArray(): number[];

    /**
     * @deprecated
     */
    multiplyVector3(vector: Vector3): any;
    multiplyVector3Array(a: any): any;
    getInverse(matrix: Matrix4, throwOnDegenerate?: boolean): Matrix3;
    flattenToArrayOffset(array: number[], offset: number): number[];
}
// ...
export class Spherical {
    constructor(radius?: number, phi?: number, theta?: number);

    radius: number;
    phi: number;
    theta: number;

    set(radius: number, phi: number, theta: number): Spherical;
    clone(): this;
    copy(other: this): this;
    makeSafe(): void;
    setFromVector3(vec3: Vector3): Spherical;
}
// ...
export interface Vector {
    setComponent(index: number, value: number): void;

    getComponent(index: number): number;

    /**
     * copy(v:T):T;
     */
    copy(v: this): this;

    /**
     * add(v:T):T;
     */
    add(v: Vector): Vector;

    /**
     * addVectors(a:T, b:T):T;
     */
    addVectors(a: Vector, b: Vector): Vector;

    /**
     * sub(v:T):T;
     */
    sub(v: Vector): Vector;

    /**
     * subVectors(a:T, b:T):T;
     */
    subVectors(a: Vector, b: Vector): Vector;

    /**
     * multiplyScalar(s:number):T;
     */
    multiplyScalar(s: number): Vector;

    /**
     * divideScalar(s:number):T;
     */
    divideScalar(s: number): Vector;

    /**
     * negate():T;
     */
    negate(): Vector;

    /**
     * dot(v:T):T;
     */
    dot(v: Vector): number;

    /**
     * lengthSq():number;
     */
    lengthSq(): number;

    /**
     * length():number;
     */
    length(): number;

    /**
     * normalize():T;
     */
    normalize(): Vector;

    /**
     * NOTE: Vector4 doesn't have the property.
     *
     * distanceTo(v:T):number;
     */
    distanceTo?(v: Vector): number;

    /**
     * NOTE: Vector4 doesn't have the property.
     *
     * distanceToSquared(v:T):number;
     */
    distanceToSquared?(v: Vector): number;

    /**
     * setLength(l:number):T;
     */
    setLength(l: number): Vector;

    /**
     * lerp(v:T, alpha:number):T;
     */
    lerp(v: Vector, alpha: number): Vector;

    /**
     * equals(v:T):boolean;
     */
    equals(v: Vector): boolean;

    /**
     * clone():T;
     */
    clone(): this;
}
// ...
export class Vector3 implements Vector {
    constructor(x?: number, y?: number, z?: number);

    x: number;
    y: number;
    z: number;

    /**
     * Sets value of this vector.
     */
    set(x: number, y: number, z: number): Vector3;

    /**
     * Sets all values of this vector.
     */
    setScalar(scalar: number): Vector3;

    /**
     * Sets x value of this vector.
     */
    setX(x: number): Vector3;

    /**
     * Sets y value of this vector.
     */
    setY(y: number): Vector3;

    /**
     * Sets z value of this vector.
     */
    setZ(z: number): Vector3;

    setComponent(index: number, value: number): void;
    getComponent(index: number): number;
    /**
     * Clones this vector.
     */
    clone(): this;
    /**
     * Copies value of v to this vector.
     */
    copy(v: this): this;

    /**
     * Adds v to this vector.
     */
    add(a: Vector3): Vector3;
    addScalar(s: number): Vector3;
    addScaledVector(v: Vector3, s: number): Vector3;

    /**
     * Sets this vector to a + b.
     */
    addVectors(a: Vector3, b: Vector3): Vector3;

    /**
     * Subtracts v from this vector.
     */
    sub(a: Vector3): Vector3;

    subScalar( s: number ): Vector3;

    /**
     * Sets this vector to a - b.
     */
    subVectors(a: Vector3, b: Vector3): Vector3;

    multiply(v: Vector3): Vector3;
    /**
     * Multiplies this vector by scalar s.
     */
    multiplyScalar(s: number): Vector3;
    multiplyVectors(a: Vector3, b: Vector3): Vector3;
    applyEuler(euler: Euler): Vector3;
    applyAxisAngle(axis: Vector3, angle: number): Vector3;
    applyMatrix3(m: Matrix3): Vector3;
    applyMatrix4(m: Matrix4): Vector3;
    applyQuaternion(q: Quaternion): Vector3;
    project(camrea: Camera): Vector3;
    unproject(camera: Camera): Vector3;
    transformDirection(m: Matrix4): Vector3;
    divide(v: Vector3): Vector3;

    /**
     * Divides this vector by scalar s.
     * Set vector to ( 0, 0, 0 ) if s == 0.
     */
    divideScalar(s: number): Vector3;
    min(v: Vector3): Vector3;
    max(v: Vector3): Vector3;
    clamp(min: Vector3, max: Vector3): Vector3;
    clampScalar(min: number, max: number): Vector3;
    clampLength(min: number, max: number): Vector3;
    floor(): Vector3;
    ceil(): Vector3;
    round(): Vector3;
    roundToZero(): Vector3;

    /**
     * Inverts this vector.
     */
    negate(): Vector3;

    /**
     * Computes dot product of this vector and v.
     */
    dot(v: Vector3): number;

    /**
     * Computes squared length of this vector.
     */
    lengthSq(): number;

    /**
     * Computes length of this vector.
     */
    length(): number;

    /**
     * Computes Manhattan length of this vector.
     * http://en.wikipedia.org/wiki/Taxicab_geometry
     */
    lengthManhattan(): number;

    /**
     * Normalizes this vector.
     */
    normalize(): Vector3;

    /**
     * Normalizes this vector and multiplies it by l.
     */
    setLength(l: number): Vector3;
    lerp(v: Vector3, alpha: number): Vector3;

    lerpVectors(v1: Vector3, v2: Vector3, alpha: number): Vector3;

    /**
     * Sets this vector to cross product of itself and v.
     */
    cross(a: Vector3): Vector3;

    /**
     * Sets this vector to cross product of a and b.
     */
    crossVectors(a: Vector3, b: Vector3): Vector3;
    projectOnVector(v: Vector3): Vector3;
    projectOnPlane(planeNormal: Vector3): Vector3;
    reflect(vector: Vector3): Vector3;
    angleTo(v: Vector3): number;

    /**
     * Computes distance of this vector to v.
     */
    distanceTo(v: Vector3): number;

    /**
     * Computes squared distance of this vector to v.
     */
    distanceToSquared(v: Vector3): number;
    distanceToManhattan(v: Vector3): number;

    setFromSpherical(s: Spherical): Vector3;
    setFromMatrixPosition(m: Matrix4): Vector3;
    setFromMatrixScale(m: Matrix4): Vector3;
    setFromMatrixColumn(matrix: Matrix4, index: number): Vector3;

    /**
     * Checks for strict equality of this vector and v.
     */
    equals(v: Vector3): boolean;

    fromArray(xyz: number[], offset?: number): Vector3;
    toArray(xyz?: number[], offset?: number): number[];
    fromBufferAttribute( attribute: BufferAttribute, index: number, offset?: number): Vector3;

    /**
     * @deprecated
     */
    getPositionFromMatrix(m: Matrix4): Vector3;
    getScaleFromMatrix(m: Matrix4): Vector3;
    getColumnFromMatrix(index: number, matrix: Matrix4): Vector3;
}
// ...
export interface TextGeometryParameters {
    font?: Font;
    size?: number;
    height?: number;
    curveSegments?: number;
    bevelEnabled?: boolean;
    bevelThickness?: number;
    bevelSize?: number;
}